apps/api/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
celery==5.3.6
redis==5.0.3
groq==0.5.0
google-generativeai==0.5.0
sentence-transformers==2.7.0
chromadb==0.4.24
PyMuPDF==1.24.1
pdfplumber==0.11.0
python-docx==1.1.0
playwright==1.42.0
beautifulsoup4==4.12.3
duckduckgo-search==5.2.1
supabase==2.4.5
spacy==3.7.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
python-multipart==0.0.9
pytz==2024.1
langgraph==0.0.38
langchain-core==0.1.45
